Separate RESTEasy specific from generic concerns
